Revert "[server][dashboard] Implement a 'Use Last Successful Prebuild' workspace creation mode"
This reverts commit 5687600aab763adc2562b31163dd337c9332aa4e.
