+++ components/dashboard/src/start/CreateWorkspace.tsx
@@ -269,9 +269,6 @@ export default class CreateWorkspace extends React.Component<CreateWorkspaceProp
             return (
                 <RunningPrebuildView
                     runningPrebuild={result.runningWorkspacePrebuild}
-                    onUseLastSuccessfulPrebuild={() =>
-                        this.createWorkspace(CreateWorkspaceMode.UseLastSuccessfulPrebuild)
-                    }
                     onIgnorePrebuild={() => this.createWorkspace(CreateWorkspaceMode.ForceNew)}
                     onPrebuildSucceeded={() => this.createWorkspace(CreateWorkspaceMode.UsePrebuild)}
                 />
@@ -534,7 +531,6 @@ interface RunningPrebuildViewProps {
         starting: RunningWorkspacePrebuildStarting;
         sameCluster: boolean;
     };
-    onUseLastSuccessfulPrebuild: () => void;
     onIgnorePrebuild: () => void;
     onPrebuildSucceeded: () => void;
 }
@@ -569,12 +565,6 @@ function RunningPrebuildView(props: RunningPrebuildViewProps) {
             {/* TODO(gpl) Copied around in Start-/CreateWorkspace. This should properly go somewhere central. */}
             <div className="h-full mt-6 w-11/12 lg:w-3/5">
                 <PrebuildLogs workspaceId={workspaceId} onIgnorePrebuild={props.onIgnorePrebuild}>
-                    <button
-                        className="secondary"
-                        onClick={() => props.onUseLastSuccessfulPrebuild && props.onUseLastSuccessfulPrebuild()}
-                    >
-                        Use Last Successful Prebuild
-                    </button>
                     <button className="secondary" onClick={() => props.onIgnorePrebuild && props.onIgnorePrebuild()}>
                         Skip Prebuild
                     </button>
