+++ components/server/ee/src/workspace/workspace-factory.ts
@@ -17,6 +17,10 @@ import {
     WorkspaceContext,
     WithSnapshot,
     WithPrebuild,
+    TaskConfig,
+    PrebuiltWorkspace,
+    WorkspaceConfig,
+    WorkspaceImageSource,
     OpenPrebuildContext,
 } from "@gitpod/gitpod-protocol";
 import { log } from "@gitpod/gitpod-protocol/lib/util/logging";
@@ -31,7 +35,6 @@ import { increasePrebuildsStartedCounter } from "../../../src/prometheus-metrics
 import { DeepPartial } from "@gitpod/gitpod-protocol/lib/util/deep-partial";
 import { EntitlementService } from "../../../src/billing/entitlement-service";
 import { getExperimentsClientForBackend } from "@gitpod/gitpod-protocol/lib/experiments/configcat-server";
-import { IncrementalPrebuildsService } from "../prebuilds/incremental-prebuilds-service";
 
 @injectable()
 export class WorkspaceFactoryEE extends WorkspaceFactory {
@@ -39,7 +42,6 @@ export class WorkspaceFactoryEE extends WorkspaceFactory {
     @inject(HostContextProvider) protected readonly hostContextProvider: HostContextProvider;
     @inject(UserCounter) protected readonly userCounter: UserCounter;
     @inject(EntitlementService) protected readonly entitlementService: EntitlementService;
-    @inject(IncrementalPrebuildsService) protected readonly incrementalPrebuildsService: IncrementalPrebuildsService;
 
     @inject(UserDB) protected readonly userDB: UserDB;
 
@@ -116,46 +118,68 @@ export class WorkspaceFactoryEE extends WorkspaceFactory {
             await assertNoPrebuildIsRunningForSameCommit();
 
             const { config } = await this.configProvider.fetchConfig({ span }, user, context.actual);
+            const imageSource = await this.imageSourceProvider.getImageSource(ctx, user, context.actual, config);
 
-            // If an incremental prebuild was requested, see if we can find a recent prebuild to act as a base.
+            // Walk back the last prebuilds and check if they are valid ancestor.
             let ws;
-            const recentPrebuild = await this.incrementalPrebuildsService.findGoodBaseForIncrementalBuild(
-                commitContext,
-                context,
-                user,
-            );
-            if (recentPrebuild) {
+            if (context.commitHistory && context.commitHistory.length > 0) {
+                // Note: This query returns only not-garbage-collected prebuilds in order to reduce cardinality
+                // (e.g., at the time of writing, the Gitpod repository has 16K+ prebuilds, but only ~300 not-garbage-collected)
+                const recentPrebuilds = await this.db
+                    .trace({ span })
+                    .findPrebuildsWithWorkpace(commitContext.repository.cloneUrl);
+
                 const loggedContext = filterForLogging(context);
-                log.info({ userId: user.id }, "Using incremental prebuild base", {
-                    basePrebuildId: recentPrebuild.id,
-                    context: loggedContext,
-                });
+                for (const recentPrebuild of recentPrebuilds) {
+                    if (
+                        !(await this.isGoodBaseforIncrementalPrebuild(
+                            context,
+                            config,
+                            imageSource,
+                            recentPrebuild.prebuild,
+                            recentPrebuild.workspace,
+                        ))
+                    ) {
+                        log.debug({ userId: user.id }, "Not using incremental prebuild base", {
+                            candidatePrebuildId: recentPrebuild.prebuild.id,
+                            context: loggedContext,
+                        });
+                        continue;
+                    }
 
-                const incrementalPrebuildContext: PrebuiltWorkspaceContext = {
-                    title: `Incremental prebuild of "${commitContext.title}"`,
-                    originalContext: commitContext,
-                    prebuiltWorkspace: recentPrebuild,
-                };
+                    log.info({ userId: user.id }, "Using incremental prebuild base", {
+                        basePrebuildId: recentPrebuild.prebuild.id,
+                        context: loggedContext,
+                    });
 
-                // repeated assertion on prebuilds triggered for same commit here, in order to
-                // reduce likelihood of duplicates if for instance handled by two different
-                // server pods.
-                await assertNoPrebuildIsRunningForSameCommit();
+                    const incrementalPrebuildContext: PrebuiltWorkspaceContext = {
+                        title: `Incremental prebuild of "${commitContext.title}"`,
+                        originalContext: commitContext,
+                        prebuiltWorkspace: recentPrebuild.prebuild,
+                    };
 
-                ws = await this.createForPrebuiltWorkspace(
-                    { span },
-                    user,
-                    incrementalPrebuildContext,
-                    normalizedContextURL,
-                );
-                // Overwrite the config from the parent prebuild:
-                //   `createForPrebuiltWorkspace` 1:1 copies the config from the parent prebuild.
-                //   Above, we've made sure that the parent's prebuild tasks (before/init/prebuild) are still the same as now.
-                //   However, other non-prebuild config items might be outdated (e.g. any command task, VS Code extension, ...)
-                //   To fix this, we overwrite the new prebuild's config with the most-recently fetched config.
-                // See also: https://github.com/gitpod-io/gitpod/issues/7475
-                //TODO(sven) doing side effects on objects back and forth is complicated and error-prone. We should rather make sure we pass in the config when creating the prebuiltWorkspace.
-                ws.config = config;
+                    // repeated assertion on prebuilds triggered for same commit here, in order to
+                    // reduce likelihood of duplicates if for instance handled by two different
+                    // server pods.
+                    await assertNoPrebuildIsRunningForSameCommit();
+
+                    ws = await this.createForPrebuiltWorkspace(
+                        { span },
+                        user,
+                        incrementalPrebuildContext,
+                        normalizedContextURL,
+                    );
+                    // Overwrite the config from the parent prebuild:
+                    //   `createForPrebuiltWorkspace` 1:1 copies the config from the parent prebuild.
+                    //   Above, we've made sure that the parent's prebuild tasks (before/init/prebuild) are still the same as now.
+                    //   However, other non-prebuild config items might be outdated (e.g. any command task, VS Code extension, ...)
+                    //   To fix this, we overwrite the new prebuild's config with the most-recently fetched config.
+                    // See also: https://github.com/gitpod-io/gitpod/issues/7475
+                    //TODO(sven) doing side effects on objects back and forth is complicated and error-prone. We should rather make sure we pass in the config when creating the prebuiltWorkspace.
+                    ws.config = config;
+
+                    break;
+                }
             }
 
             // repeated assertion on prebuilds triggered for same commit here, in order to
@@ -211,6 +235,85 @@ export class WorkspaceFactoryEE extends WorkspaceFactory {
         } finally {
             span.finish();
         }
+    }
+
+    private async isGoodBaseforIncrementalPrebuild(
+        context: StartPrebuildContext,
+        config: WorkspaceConfig,
+        imageSource: WorkspaceImageSource,
+        candidatePrebuild: PrebuiltWorkspace,
+        candidate: Workspace,
+    ): Promise<boolean> {
+        if (!context.commitHistory || context.commitHistory.length === 0) {
+            return false;
+        }
+        if (!CommitContext.is(candidate.context)) {
+            return false;
+        }
+
+        // we are only considering available prebuilds
+        if (candidatePrebuild.state !== "available") {
+            return false;
+        }
+
+        // we are only considering full prebuilds
+        if (!!candidate.basedOnPrebuildId) {
+            return false;
+        }
+
+        const candidateCtx = candidate.context;
+        if (
+            candidateCtx.additionalRepositoryCheckoutInfo?.length !==
+            context.additionalRepositoryCommitHistories?.length
+        ) {
+            // different number of repos
+            return false;
+        }
+
+        if (!context.commitHistory.some((sha) => sha === candidateCtx.revision)) {
+            return false;
+        }
+
+        // check the commits are included in the commit history
+        for (const subRepo of candidateCtx.additionalRepositoryCheckoutInfo || []) {
+            const matchIngRepo = context.additionalRepositoryCommitHistories?.find(
+                (repo) => repo.cloneUrl === subRepo.repository.cloneUrl,
+            );
+            if (!matchIngRepo || !matchIngRepo.commitHistory.some((sha) => sha === subRepo.revision)) {
+                return false;
+            }
+        }
+
+        // ensure the image source hasn't changed (skips older images)
+        if (JSON.stringify(imageSource) !== JSON.stringify(candidate.imageSource)) {
+            log.debug(`Skipping parent prebuild: Outdated image`, {
+                imageSource,
+                parentImageSource: candidate.imageSource,
+            });
+            return false;
+        }
+
+        // ensure the tasks haven't changed
+        const filterPrebuildTasks = (tasks: TaskConfig[] = []) =>
+            tasks
+                .map((task) =>
+                    Object.keys(task)
+                        .filter((key) => ["before", "init", "prebuild"].includes(key))
+                        // @ts-ignore
+                        .reduce((obj, key) => ({ ...obj, [key]: task[key] }), {}),
+                )
+                .filter((task) => Object.keys(task).length > 0);
+        const prebuildTasks = filterPrebuildTasks(config.tasks);
+        const parentPrebuildTasks = filterPrebuildTasks(candidate.config.tasks);
+        if (JSON.stringify(prebuildTasks) !== JSON.stringify(parentPrebuildTasks)) {
+            log.debug(`Skipping parent prebuild: Outdated prebuild tasks`, {
+                prebuildTasks,
+                parentPrebuildTasks,
+            });
+            return false;
+        }
+
+        return true;
     }
 
     protected async createForPrebuiltWorkspace(
